Changed DatabaseUtil (static) into instantiable database manager
This allows us to use the manager in different test harnesses.
