+++ src/test/java/net/krotscheck/api/admin/v1/user/resource/UserServiceTest.java
@@ -24,12 +24,10 @@ import net.krotscheck.features.exception.ExceptionFeature;
 import net.krotscheck.features.jackson.JacksonFeature;
 import net.krotscheck.jersey2.hibernate.context.SearchIndexContextListener;
 import net.krotscheck.test.ContainerTest;
-import net.krotscheck.test.DatabaseUtil;
 import net.krotscheck.util.ResourceUtil;
 import org.apache.http.HttpStatus;
 import org.glassfish.jersey.server.ResourceConfig;
 import org.glassfish.jersey.test.TestProperties;
-import org.junit.AfterClass;
 import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.Test;
@@ -82,18 +80,7 @@ public final class UserServiceTest extends ContainerTest {
     @BeforeClass
     public static void setupTestData() {
         // Load data
-        DatabaseUtil.loadTestData(TEST_DATA);
-        // Rebuild the search index.
-        new SearchIndexContextListener()
-                .contextInitialized(mock(ServletContextEvent.class));
-    }
-
-    /**
-     * Clean test data.
-     */
-    @AfterClass
-    public static void cleanTestData() {
-        DatabaseUtil.clearTestData(TEST_DATA);
+        loadTestData(TEST_DATA);
     }
 
     /**
